refactor: 인증 관련 로직 member -> auth 로 위치 변경
