+++ src/main/java/com/example/sinitto/auth/dto/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package com.example.sinitto.member.dto;
+package com.example.sinitto.auth.dto;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
