+++ src/main/java/com/example/sinitto/auth/dto/KakaoUserResponse.java
@@ -1,4 +1,4 @@
-package com.example.sinitto.member.dto;
+package com.example.sinitto.auth.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
