+++ src/main/java/com/example/sinitto/auth/dto/RegisterResponse.java
@@ -1,0 +1,8 @@
+package com.example.sinitto.auth.dto;
+
+public record RegisterResponse(
+        String accessToken,
+        String refreshToken,
+        boolean isSinitto
+) {
+}
