+++ src/main/java/com/example/sinitto/auth/exception/NotUniqueException.java
@@ -1,0 +1,7 @@
+package com.example.sinitto.auth.exception;
+
+public class NotUniqueException extends RuntimeException {
+    public NotUniqueException(String message) {
+        super(message);
+    }
+}
