+++ src/main/java/com/example/sinitto/auth/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.example.sinitto.member.exception;
+package com.example.sinitto.auth.exception;
 
 public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
