--- src/main/java/com/example/sinitto/member/dto/GuardSignupRequest.java
@@ -1,7 +1,0 @@
-package com.example.sinitto.member.dto;
-
-public record GuardSignupRequest(
-        String name,
-        String email,
-        String phoneNumber) {
-}
