--- src/main/java/com/example/sinitto/member/dto/RegisterResponse.java
@@ -1,7 +1,0 @@
-package com.example.sinitto.member.dto;
-
-public record RegisterResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
