+++ src/main/java/com/example/sinitto/member/entity/Member.java
@@ -34,4 +34,8 @@ public class Member {
     public Long getId() {
         return id;
     }
+
+    public boolean isSinitto() {
+        return isSinitto;
+    }
 }
