--- src/main/java/com/example/sinitto/member/exception/NotUniqueException.java
@@ -1,7 +1,0 @@
-package com.example.sinitto.member.exception;
-
-public class NotUniqueException extends RuntimeException{
-    public NotUniqueException(String message) {
-        super(message);
-    }
-}
